feat: Adding settings, utils, and writer close
